include sad path testing for creating a new bulk discount for a merchant
