+++ app/controllers/merchants/bulk_discounts_controller.rb
@@ -16,8 +16,13 @@ class Merchants::BulkDiscountsController < ApplicationController
 
   def create 
     merchant 
-    bulk_discount = merchant.bulk_discounts.create(bk_params)
-    redirect_to merchant_bulk_discounts_path(merchant)
+    bulk_discount = merchant.bulk_discounts.new(bk_params)
+    if bulk_discount.save(bk_params)
+      redirect_to merchant_bulk_discounts_path(merchant)
+    else 
+      flash[:alert] = bulk_discount.errors.full_messages.to_sentence 
+      redirect_to new_merchant_bulk_discount_path(merchant)
+    end 
   end
 
   private
