+++ spec/features/merchants/bulk_discounts/index_spec.rb
@@ -45,7 +45,12 @@ RSpec.describe 'Bulk Discount Index Page' do
 
       within "#bk_discount_#{bk_1.id}" do 
         expect(page).to have_link "#{bk_1.id}", href: merchant_bulk_discount_path(merchant_1, bk_1)
+        click_link bk_1.id
       end
+
+      expect(current_path).to eq merchant_bulk_discount_path(merchant_1, bk_1)
+
+      visit merchant_bulk_discounts_path(merchant_1)
 
       within "#bk_discount_#{bk_2.id}" do 
         expect(page).to have_link "#{bk_2.id}", href: merchant_bulk_discount_path(merchant_1, bk_2)
@@ -55,36 +60,5 @@ RSpec.describe 'Bulk Discount Index Page' do
         expect(page).to have_link "#{bk_3.id}", href: merchant_bulk_discount_path(merchant_1, bk_3)
       end
     end
-
-    describe 'create a new discount' do 
-      it 'has a link to add a new bulk discount' do 
-        visit merchant_bulk_discounts_path(merchant_1)
-
-        click_link "Create a New Bulk Discount"
-
-        expect(current_path).to eq new_merchant_bulk_discount_path(merchant_1)
-      end
-
-      it 'adds a new bulk discount to the bulk discount index page' do 
-        visit new_merchant_bulk_discount_path(merchant_1)
-
-        fill_in 'Discount', with: 25
-        fill_in 'Quantity threshold', with: 5
-        click_on 'Create New Discount'
-
-        new_bk = BulkDiscount.last 
-        expect(current_path).to eq merchant_bulk_discounts_path(merchant_1)
-
-        within "#bk_discount_#{new_bk.id}" do 
-          expect(page).to have_content(new_bk.id)
-          expect(page).to have_content("Percent Discount: %#{new_bk.discount}")
-          expect(page).to have_content("Quantity Threshold: #{new_bk.quantity_threshold} items")    
-        end
-      end
-
-      xit 'test for sad path.....' do 
-
-      end
-    end
-  end
-end
+  end 
+end
